Trivia winners are platform aware
