+++ src/variables/unorganized.ts
@@ -97,76 +97,6 @@ export const mageTriviaPossibleAnswers: ReplaceVariable = {
     }
 };
 
-export const mageTriviaGameResultsRaw: ReplaceVariable = {
-    definition: {
-        handle: "mageTriviaGameResultsRaw",
-        description: "Returns the raw object containing the results of the last trivia game played. (An empty object is returned if the game is not completed.)",
-        possibleDataOutput: ["object"]
-    },
-    evaluator: async () => {
-        const lastGameResults = triviaGame.getGameManager().getGameState();
-        if (!lastGameResults) {
-            logger('warn', 'Called mageTriviaGameResultsRaw variable when the game state was not initialized.');
-            return {};
-        }
-        if (!lastGameResults.complete) {
-            logger('warn', 'Called mageTriviaGameResultsRaw variable when the game was not completed.');
-            return {};
-        }
-
-        logger('debug', `mageTriviaGameResultsRaw: ${JSON.stringify(lastGameResults)}`);
-        return lastGameResults;
-    }
-};
-
-export const mageTriviaWinners: ReplaceVariable = {
-    definition: {
-        handle: "mageTriviaWinners",
-        description: "Returns an array of the winners of the last trivia game played. (An empty array is returned if the game is not completed.)",
-        possibleDataOutput: ["array"]
-    },
-    evaluator: async () => {
-        const lastGameResults = triviaGame.getGameManager().getGameState();
-        if (!lastGameResults) {
-            logger('warn', 'Called mageTriviaWinners variable when no last game results were available.');
-            return [];
-        }
-        if (!lastGameResults.complete) {
-            logger('warn', 'Called mageTriviaWinners variable when the game was not completed.');
-            return [];
-        }
-        const winners = lastGameResults.winners.map((winner) => {
-            return winner.userDisplayName;
-        });
-        logger('debug', `mageTriviaWinners: ${JSON.stringify(winners)}`);
-        return winners;
-    }
-};
-
-export const mageTriviaWinnersWithPoints: ReplaceVariable = {
-    definition: {
-        handle: "mageTriviaWinnersWithPoints",
-        description: "Returns an array of the winners of the last trivia game played with their points. (An empty array is returned if the game is not completed.)",
-        possibleDataOutput: ["array"]
-    },
-    evaluator: async () => {
-        const lastGameResults = triviaGame.getGameManager().getGameState();
-        if (!lastGameResults) {
-            logger('warn', 'Called mageTriviaLastGameWinnersWithPoints variable when no last game results were available.');
-            return [];
-        }
-        if (!lastGameResults.complete) {
-            logger('warn', 'Called mageTriviaLastGameWinnersWithPoints variable when the game was not completed.');
-            return [];
-        }
-        const winnersWithPoints = lastGameResults.winners.map((winner) => {
-            return `${winner.userDisplayName} (+${winner.points})`;
-        });
-        logger('debug', `mageTriviaLastGameWinnersWithPoints: ${JSON.stringify(winnersWithPoints)}`);
-        return winnersWithPoints;
-    }
-};
-
 function getAskedQuestion(triviaGame: TriviaGame): askedQuestion | undefined {
     const game = triviaGame.getGameManager().getQuestion();
     if (game) {
